Split :help command to own module
This will allow defining a bunch of helper functions for beautification
here without spamming the _modules.
